# src/frontend/src/lib/api/openapi.json
{
  "openapi" : "3.1.0",
  "components" : {
    "schemas" : {
...
      "BucketVoteRequest" : {
        "type" : "object",
        "required" : [ "bucketId" ],
        "properties" : {
          "bucketId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "isSample" : {
            "type" : "boolean"
          },
          "minEffort" : {
            "type" : "number",
            "format" : "double"
          },
          "expectedEffort" : {
            "type" : "number",
            "format" : "double"
          },
          "maxEffort" : {
            "type" : "number",
            "format" : "double"
          }
        }
      },
...
      "CreateSessionRequest" : {
        "type" : "object",
        "required" : [ "estimationId", "title", "itemLogicalIds" ],
        "properties" : {
          "estimationId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "title" : {
            "type" : "string"
          },
          "itemLogicalIds" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "moderatorEstimates" : {
            "type" : "boolean"
          }
        }
      },
...
      "NotesRequest" : {
        "type" : "object",
        "required" : [ "notes" ],
        "properties" : {
          "notes" : {
            "type" : "string"
          }
        }
      },
...
      "ParticipantDto" : {
        "type" : "object",
        "required" : [ "subjectId", "displayName", "role", "agreed" ],
        "properties" : {
          "subjectId" : {
            "type" : "string"
          },
          "displayName" : {
            "type" : [ "string", "null" ]
          },
          "role" : {
            "$ref" : "#/components/schemas/ParticipantRole"
          },
          "agreed" : {
            "type" : "boolean"
          }
        }
      },
      "ParticipantRole" : {
        "type" : "string",
        "enum" : [ "MODERATOR", "ESTIMATOR" ]
      },
...
      "SessionDto" : {
        "type" : "object",
        "required" : [ "id", "title", "estimationId", "status", "moderatorSubjectId", "currentItemIndex", "currentPhase", "moderatorEstimates", "items", "participants" ],
        "properties" : {
          "id" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "title" : {
            "type" : "string"
          },
          "estimationId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "status" : {
            "$ref" : "#/components/schemas/SessionStatus"
          },
          "moderatorSubjectId" : {
            "type" : "string"
          },
          "currentItemIndex" : {
            "type" : "integer",
            "format" : "int32"
          },
          "currentPhase" : {
            "$ref" : "#/components/schemas/SessionPhase"
          },
          "moderatorEstimates" : {
            "type" : "boolean"
          },
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SessionItemDto"
            }
          },
          "participants" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ParticipantDto"
            }
          }
        }
      },
...
      "SessionPhase" : {
        "type" : "string",
        "enum" : [ "PHASE1", "PHASE2" ]
      },
      "SessionStatus" : {
        "type" : "string",
        "enum" : [ "CREATED", "RUNNING", "SUSPENDED", "FINALIZED", "ENDED_EARLY", "CANCELLED" ]
      },
...
      "UUID" : {
        "type" : "string",
        "format" : "uuid",
        "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      },
...
      "VoteRequest" : {
        "type" : "object",
        "required" : [ "minEffort", "expectedEffort", "maxEffort" ],
        "properties" : {
          "minEffort" : {
            "type" : "number",
            "format" : "double"
          },
          "expectedEffort" : {
            "type" : "number",
            "format" : "double"
          },
          "maxEffort" : {
            "type" : "number",
            "format" : "double"
          }
        }
      },
      "WsTicketDto" : {
        "type" : "object",
        "required" : [ "ticket" ],
        "properties" : {
          "ticket" : {
            "type" : "string"
          }
        }
      }
    },
    "securitySchemes" : {
      "SecurityScheme" : {
        "description" : "Authentication",
        "type" : "openIdConnect",
        "openIdConnectUrl" : "https://login.microsoftonline.com/common/v2.0/.well-known/openid-configuration"
      }
    }
  },
...
  "paths" : {
...
    "/api/sessions" : {
      "get" : {
        "summary" : "List sessions — of one estimation when estimationId is given, otherwise all joinable (CREATED/RUNNING) sessions",
        "tags" : [ "Estimation sessions" ],
        "parameters" : [ {
          "description" : "Filter to one estimation; omit for all joinable sessions",
          "required" : false,
          "name" : "estimationId",
          "in" : "query",
          "schema" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/UUID"
            }, {
              "type" : "null"
            } ]
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/SessionDto"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "ESTIMATOR" ]
        } ]
      },
      "post" : {
        "summary" : "Create a session (the current user becomes the moderator)",
        "tags" : [ "Estimation sessions" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateSessionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Session created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SessionDto"
                }
              }
            }
          },
          "400" : {
            "description" : "A logicalId is not a draft leaf, or the estimation has no draft"
          },
          "403" : {
            "description" : "Not an ESTIMATOR"
          },
          "401" : {
            "description" : "Not Authorized"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "ESTIMATOR" ]
        } ]
      }
    },
    "/api/sessions/{id}" : {
      "get" : {
        "summary" : "Read a session (blind items expose only a vote count)",
        "tags" : [ "Estimation sessions" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SessionDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Session not found"
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "ESTIMATOR" ]
        } ]
      }
    },
    "/api/sessions/{id}/agree" : {
      "post" : {
        "summary" : "Estimator: mark agreement on the current item",
        "tags" : [ "Estimation sessions" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SessionDto"
                }
              }
            }
          },
          "409" : {
            "description" : "Not a participant"
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "ESTIMATOR" ]
        } ]
      }
    },
    "/api/sessions/{id}/cancel" : {
      "post" : {
        "summary" : "Moderator: cancel the session",
        "tags" : [ "Estimation sessions" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SessionDto"
                }
              }
            }
          },
          "403" : {
            "description" : "Not the moderator"
          },
          "401" : {
            "description" : "Not Authorized"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "ESTIMATOR" ]
        } ]
      }
    },
    "/api/sessions/{id}/end-early" : {
      "post" : {
        "summary" : "Moderator: end the session before its last item, keeping the results so far",
        "tags" : [ "Estimation sessions" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SessionDto"
                }
              }
            }
          },
          "403" : {
            "description" : "Not the moderator"
          },
          "404" : {
            "description" : "Session not found"
          },
          "409" : {
            "description" : "Session is not in a state that allows this transition"
          },
          "401" : {
            "description" : "Not Authorized"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "ESTIMATOR" ]
        } ]
      }
    },
    "/api/sessions/{id}/items/current/finalize" : {
      "post" : {
        "summary" : "Moderator: finalize the current item (writes back to the draft) and advance",
        "tags" : [ "Estimation sessions" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SessionDto"
                }
              }
            }
          },
          "403" : {
            "description" : "Not the moderator"
          },
          "409" : {
            "description" : "Current item is not in PHASE2, or has no votes"
          },
          "401" : {
            "description" : "Not Authorized"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "ESTIMATOR" ]
        } ]
      }
    },
    "/api/sessions/{id}/items/current/notes" : {
      "put" : {
        "summary" : "Moderator: set the current item's discussion notes",
        "tags" : [ "Estimation sessions" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NotesRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SessionDto"
                }
              }
            }
          },
          "403" : {
            "description" : "Not the moderator"
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "ESTIMATOR" ]
        } ]
      }
    },
    "/api/sessions/{id}/items/current/phase2" : {
      "post" : {
        "summary" : "Moderator: reveal — advance the current item to PHASE2",
        "tags" : [ "Estimation sessions" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SessionDto"
                }
              }
            }
          },
          "403" : {
            "description" : "Not the moderator"
          },
          "409" : {
            "description" : "Current item is not in PHASE1"
          },
          "401" : {
            "description" : "Not Authorized"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "ESTIMATOR" ]
        } ]
      }
    },
    "/api/sessions/{id}/join" : {
      "post" : {
        "summary" : "Join the current user as an ESTIMATOR participant (idempotent)",
        "tags" : [ "Estimation sessions" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SessionDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Session not found"
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "ESTIMATOR" ]
        } ]
      }
    },
    "/api/sessions/{id}/resume" : {
      "post" : {
        "summary" : "Moderator: resume a paused session (SUSPENDED → RUNNING)",
        "tags" : [ "Estimation sessions" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SessionDto"
                }
              }
            }
          },
          "403" : {
            "description" : "Not the moderator"
          },
          "404" : {
            "description" : "Session not found"
          },
          "409" : {
            "description" : "Session is not in a state that allows this transition"
          },
          "401" : {
            "description" : "Not Authorized"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "ESTIMATOR" ]
        } ]
      }
    },
    "/api/sessions/{id}/start" : {
      "post" : {
        "summary" : "Moderator: start the session (CREATED → RUNNING)",
        "tags" : [ "Estimation sessions" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SessionDto"
                }
              }
            }
          },
          "403" : {
            "description" : "Not the moderator"
          },
          "409" : {
            "description" : "Session is not in CREATED"
          },
          "401" : {
            "description" : "Not Authorized"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "ESTIMATOR" ]
        } ]
      }
    },
    "/api/sessions/{id}/suspend" : {
      "post" : {
        "summary" : "Moderator: pause the session (RUNNING → SUSPENDED), keeping all progress",
        "tags" : [ "Estimation sessions" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SessionDto"
                }
              }
            }
          },
          "403" : {
            "description" : "Not the moderator"
          },
          "404" : {
            "description" : "Session not found"
          },
          "409" : {
            "description" : "Session is not in a state that allows this transition"
          },
          "401" : {
            "description" : "Not Authorized"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "ESTIMATOR" ]
        } ]
      }
    },
    "/api/sessions/{id}/votes" : {
      "post" : {
        "summary" : "Estimator: upsert a vote for the current item + current phase",
        "tags" : [ "Estimation sessions" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/VoteRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SessionDto"
                }
              }
            }
          },
          "409" : {
            "description" : "Not a participant, session not RUNNING, or no current item"
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "ESTIMATOR" ]
        } ]
      }
    },
    "/api/sessions/{id}/votes/bucket" : {
      "post" : {
        "summary" : "Estimator: upsert a BUCKET vote for the current item + current phase",
        "tags" : [ "Estimation sessions" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BucketVoteRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SessionDto"
                }
              }
            }
          },
          "409" : {
            "description" : "Not a participant, session not RUNNING, no current item, the estimation is not bucket+sampled, or the bucket is not one of its buckets"
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "ESTIMATOR" ]
        } ]
      }
    },
    "/api/sessions/{id}/ws-ticket" : {
      "post" : {
        "summary" : "Issue a short-lived single-use WebSocket handshake ticket for the current participant",
        "tags" : [ "Estimation sessions" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WsTicketDto"
                }
              }
            }
          },
          "403" : {
            "description" : "Not a participant of the session"
          },
          "404" : {
            "description" : "Session not found"
          },
          "401" : {
            "description" : "Not Authorized"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "ESTIMATOR" ]
        } ]
      }
    },
...
  },
  "info" : {
    "title" : "Pythia API",
    "version" : "1.0.0-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}